feat: rename old Chat page to Discussion and move to marker popup; add Chat page mockup
