+++ mobile_app/lib/pages/pages.dart
@@ -1,7 +1,8 @@
 export 'primary_page.dart';
 export 'menu_page.dart';
-export 'chat_page.dart';
+export 'discussion_page.dart';
 export 'catch_page.dart';
 export 'profile_page.dart';
 export 'settings_page.dart';
 export 'about_page.dart';
+export 'chat_page.dart';
